requirements.txt
Django==3.2.6
djangorestframework==3.12.4
pytest==8.0.0
pytest-django==4.8.0